MINOR: Nest tiered storage test plan in builder
